requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosmtplib==3.0.2
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.1.0
backoff==2.2.1
bcrypt==4.0.1
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
environs==14.1.0
exceptiongroup==1.2.2
fastapi==0.115.7
fastapi-cli==0.0.7
fastapi-mail==1.4.2
fastapi-pagination==0.12.34
flower==2.0.1
fqdn==1.5.1
frozenlist==1.5.0
graphql-core==3.2.6
greenlet==3.1.1
h11==0.14.0
harfile==0.3.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanize==4.11.0
hypothesis==6.124.7
hypothesis-graphql==0.11.1
hypothesis-jsonschema==0.23.1
idna==3.10
iniconfig==2.0.0
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
junit-xml==1.9
kombu==5.4.2
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
mdurl==0.1.2
multidict==6.1.0
numpy==2.2.5
orjson==3.10.15
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyrate-limiter==3.7.0
pytest==8.3.4
pytest-subtests==0.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.5.1
python-multipart==0.0.20
python-socketio==5.8.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
ruff==0.9.3
schemathesis==3.39.8
scikit-learn==1.6.1
scipy==1.15.2
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.37
sqlmodel==0.0.22
starlette==0.45.3
starlette-testclient==0.4.1
threadpoolctl==3.6.0
tomli==2.2.1
tomli_w==1.2.0
tornado==6.4.2
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
ujson==5.10.0
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
